Nuke the static
